backend/fix : default plan allocation in ysvehicle category based
